refactor: Hero 개별 적용을 위해 수정
