+++ src/app/consulting/transfer/layout.tsx
@@ -1,46 +1,17 @@
-'use client';
-
-import React, { useEffect, useState } from 'react';
-import { useMediaQuery, useTheme } from '@mui/material';
-import Box from '@mui/material/Box';
-import {
-  LARGE_LAYOUT_WIDTH,
-  MEDIUM_LAYOUT_WIDTH,
-  SMALL_LAYOUT_WIDTH,
-} from '@/components/layout/general-layout/constants';
+import React from 'react';
+import { Metadata } from 'next';
 
 interface TransferLayoutProps {
   children: React.ReactNode;
 }
+
+export const metadata: Metadata = {
+  title: '코어창업 | 양도 컨설팅',
+  description: '코어창업 양도 컨설팅 신청',
+};
+
 const TransferLayout = ({ children }: TransferLayoutProps) => {
-  const theme = useTheme();
-  const isUpLarge = useMediaQuery(theme.breakpoints.up('lg'));
-  const isMedium = useMediaQuery(theme.breakpoints.only('md'));
-  const [layoutWidth, setLayoutWidth] = useState(LARGE_LAYOUT_WIDTH);
-
-  useEffect(() => {
-    if (isUpLarge) {
-      setLayoutWidth(LARGE_LAYOUT_WIDTH);
-      return;
-    }
-
-    if (isMedium) {
-      setLayoutWidth(MEDIUM_LAYOUT_WIDTH);
-      return;
-    }
-
-    setLayoutWidth(SMALL_LAYOUT_WIDTH);
-  }, [isUpLarge, isMedium]);
-
-  return (
-    <Box
-      sx={{
-        width: layoutWidth,
-      }}
-    >
-      {children}
-    </Box>
-  );
+  return children;
 };
 
 export default TransferLayout;
