+++ src/app/partnership/page.tsx
@@ -1,5 +1,6 @@
 import Box from '@mui/material/Box';
 import { Metadata } from 'next';
+import GeneralHero from '@/components/layout/general-layout/GeneralHero';
 
 export const metadata: Metadata = {
   title: '코어창업 | 협업 신청',
@@ -8,18 +9,9 @@ export const metadata: Metadata = {
 
 export default function Partnership() {
   return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'column',
-        padding: 5,
-        alignItems: 'flex-start',
-        border: '1px solid',
-        borderColor: 'divider',
-        borderRadius: 2,
-      }}
-    >
-      내용 확인 필요함
-    </Box>
+    <>
+      <GeneralHero />
+      <Box>내용 확인 필요함</Box>
+    </>
   );
 }
