+++ src/components/layout/general-layout/GeneralHero.tsx
@@ -5,9 +5,7 @@ import { useTheme, useMediaQuery } from '@mui/material';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Image from 'next/image';
-import { usePathname } from 'next/navigation';
 import { OFF_WHITE_COLOR } from '@/constants/color';
-import { CorePointRoutes } from '@/constants/routes';
 import type { TypographyVariant } from '@mui/material';
 
 const LARGE_HERO_HEIGHT = '500px';
@@ -19,18 +17,17 @@ const MEDIUM_TITLE_SIZE = 'h4';
 const SMALL_TITLE_SIZE = 'h5';
 
 interface GeneralHeroProps {
+  title?: string;
+  description?: string;
+  imgSrc?: string;
   heroComponent?: React.ReactNode;
 }
 const GeneralHero = (props: GeneralHeroProps) => {
-  const { heroComponent } = props;
-  const pathname = usePathname();
+  const { heroComponent, description, imgSrc, title } = props;
   const theme = useTheme();
   const isUpLarge = useMediaQuery(theme.breakpoints.up('lg'));
   const isMedium = useMediaQuery(theme.breakpoints.only('md'));
   const [heroHeight, setHeroHeight] = useState(LARGE_HERO_HEIGHT);
-  const [imgSrc, setImgSrc] = useState('');
-  const [title, setTitle] = useState('');
-  const [description, setDescription] = useState('');
   const [titleSize, setTitleSize] =
     useState<TypographyVariant>(LARGE_TITLE_SIZE);
 
@@ -50,50 +47,6 @@ const GeneralHero = (props: GeneralHeroProps) => {
     setHeroHeight(SMALL_HERO_HEIGHT);
     setTitleSize(SMALL_TITLE_SIZE);
   }, [isUpLarge, isMedium]);
-
-  useEffect(() => {
-    if (!pathname) return;
-
-    switch (pathname) {
-      case CorePointRoutes.HOME:
-        setTitle('코어 창업');
-        setDescription(
-          '코어창업이 창업의 문을 여는 열쇠가 되어 드리겠습니다. (스타일 수정 예정)',
-        );
-        setImgSrc('https://source.unsplash.com/random?main');
-        break;
-      case CorePointRoutes.STORE:
-        setTitle('매물 정보');
-        setDescription(
-          'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Cumque, eum eos? Animi obcaecati neque odio facere autem sapiente fugit, minima, quae ratione rerum omnis ducimus exercitationem facilis, magnam amet similique?',
-        );
-        setImgSrc('https://source.unsplash.com/random?store');
-        break;
-      case CorePointRoutes.OPENING_CONSULTING:
-        setTitle('창업컨설팅 신청');
-        setDescription(
-          'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Cumque, eum eos? Animi obcaecati neque odio facere autem sapiente fugit, minima, quae ratione rerum omnis ducimus exercitationem facilis, magnam amet similique?',
-        );
-        setImgSrc('https://source.unsplash.com/random?contract');
-        break;
-      case CorePointRoutes.TRANSFER_CONSULTING:
-        setTitle('양도컨설팅 신청');
-        setDescription(
-          'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Cumque, eum eos? Animi obcaecati neque odio facere autem sapiente fugit, minima, quae ratione rerum omnis ducimus exercitationem facilis, magnam amet similique?',
-        );
-        setImgSrc('https://source.unsplash.com/random?sell');
-        break;
-      case CorePointRoutes.PARTNERSHIP:
-        setTitle('협업 신청');
-        setDescription(
-          'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Cumque, eum eos? Animi obcaecati neque odio facere autem sapiente fugit, minima, quae ratione rerum omnis ducimus exercitationem facilis, magnam amet similique?',
-        );
-        setImgSrc('https://source.unsplash.com/random?partnership');
-        break;
-      default:
-        break;
-    }
-  }, [pathname]);
 
   return (
     <Box sx={{ width: '100%', height: heroHeight, position: 'relative' }}>
