+++ src/components/layout/general-layout/GeneralLayout.tsx
@@ -1,24 +1,15 @@
-'use client';
-
 import React from 'react';
-import { useTheme } from '@mui/material';
 import Box from '@mui/material/Box';
-import { usePathname } from 'next/navigation';
 import FabGroup from '@/components/common/fab-group/FabGroup';
 import { OFF_WHITE_COLOR } from '@/constants/color';
-import { CorePointRoutes } from '@/constants/routes';
 import GeneralFooter from './GeneralFooter';
 import GeneralHeader from './GeneralHeader';
-import GeneralHero from './GeneralHero';
 
 interface GeneralLayoutInterface {
   children: React.ReactNode;
 }
 const GeneralLayout = (props: GeneralLayoutInterface) => {
   const { children } = props;
-  const theme = useTheme();
-  const pathname = usePathname();
-  const disableHero = pathname.includes(CorePointRoutes.STORE + '/');
 
   return (
     <Box
@@ -33,15 +24,14 @@ const GeneralLayout = (props: GeneralLayoutInterface) => {
       }}
     >
       <GeneralHeader />
-      {!disableHero && <GeneralHero />}
       <FabGroup />
       <Box
         sx={{
           flexGrow: 1,
           display: 'flex',
+          flexDirection: 'column',
           width: '100%',
-          justifyContent: 'center',
-          [theme.breakpoints.down('lg')]: { padding: 3 },
+          alignItems: 'center',
         }}
       >
         {children}
